\name{ooc-package}
\alias{ooc-package}
\docType{package}
\title{Ordered Optimal Classification}
\description{Performs Ordered Optimal Classification, an extension of Poole's (2000) nonparametric unfolding procedure for the analysis of ordinal choice data (e.g., ``Strongly Agree,'' ``Somewhat Agree,'' ``Somewhat Disagree,'' ``Strongly Disagree'').}
\details{
\tabular{ll}{
Package: \tab ooc\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-10-13\cr
License: \tab GPL-2 \cr
}}
\author{Christopher Hare and Keith T. Poole}
\references{Poole, Keith T. 2000. ``Nonparametric Unfolding of Binary Choice Data.'' \emph{Political Analysis} 8(3): 211--237.}
\keyword{ooc}
\seealso{oc}


